package: add new npm scripts, use esbuild to minify
This adjusts the build scripts to add the new scripts required to build.
The new additions are:
- esbuild-base: the core esbuild command for bundling the extension
- esbuild-watch: the watch version used for development
- package: package only step
- dev-package: package the extension with the development version id

This updates the `vscode:prepublish` script to minify the extension
prior to building the extension package.

`createDevPackage` is left as a compatibility option, however, it does
not work on all platforms (i.e. Windows) due to the use of bash specific
scripting.